# openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot Backend API
  description: Production-ready chatbot backend with Fastify and TypeScript
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Returns the current health status of the API
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                timestamp: "2025-09-02T11:18:17.355Z"
                version: "1.0.0"

  /api/answer:
    post:
      tags:
        - Answer
      summary: Get an answer for a question
      description: Get an answer for a question using AI/heuristic search
      operationId: getAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
            example:
              question: "Wie kann ich meine Rechnung einsehen?"
      responses:
        '200':
          description: Answer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
              example:
                answer: "Sie können Ihre Rechnung online in Ihrem Kundenkonto unter 'Meine Rechnungen' einsehen. Alternativ erhalten Sie diese auch per E-Mail."
                confidence: 0.8
                sourceId: "1"
                timestamp: "2025-09-02T11:18:17.355Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "ValidationError"
                message: "body/question must NOT have fewer than 3 characters"
                details:
                  - field: "question"
                    message: "must NOT have fewer than 3 characters"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "InternalServerError"
                message: "An unexpected error occurred"

  /api/feedback:
    post:
      tags:
        - Feedback
      summary: Submit feedback for an answer
      description: Submit feedback indicating whether an answer was helpful or not
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
            example:
              question: "Wie kann ich meine Rechnung einsehen?"
              helpful: true
              sourceId: "1"
      responses:
        '200':
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
              example:
                ok: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "ValidationError"
                message: "body must have required property 'question'"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "InternalServerError"
                message: "An unexpected error occurred"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - version
      properties:
        status:
          type: string
          enum: [ok]
          description: Health status of the service
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp of the health check
        version:
          type: string
          description: Version of the API
      example:
        status: ok
        timestamp: "2025-09-02T11:18:17.355Z"
        version: "1.0.0"

    AnswerRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 3
          maxLength: 1000
          description: The question to get an answer for (minimum 3 characters)
      example:
        question: "Wie kann ich meine Rechnung einsehen?"

    AnswerResponse:
      type: object
      required:
        - answer
        - confidence
        - timestamp
      properties:
        answer:
          type: string
          description: The answer to the question
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score of the answer (0-1)
        sourceId:
          type: string
          description: ID of the source FAQ (if found)
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp of when the answer was generated
      example:
        answer: "Sie können Ihre Rechnung online in Ihrem Kundenkonto unter 'Meine Rechnungen' einsehen. Alternativ erhalten Sie diese auch per E-Mail."
        confidence: 0.8
        sourceId: "1"
        timestamp: "2025-09-02T11:18:17.355Z"

    FeedbackRequest:
      type: object
      required:
        - question
        - helpful
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 1000
          description: The original question that was asked
        helpful:
          type: boolean
          description: Whether the answer was helpful or not
        sourceId:
          type: string
          description: Optional ID of the FAQ that provided the answer
      example:
        question: "Wie kann ich meine Rechnung einsehen?"
        helpful: true
        sourceId: "1"

    FeedbackResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          enum: [true]
          description: Success indicator
      example:
        ok: true

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Specific error message for the field
          description: Detailed validation errors (optional)
      example:
        error: "ValidationError"
        message: "Invalid request data"
        details:
          - field: "question"
            message: "must NOT have fewer than 3 characters"

tags:
  - name: Health
    description: Health check operations
  - name: Answer
    description: Answer retrieval operations  
  - name: Feedback
    description: Feedback submission operations